Final Version/requirements.txt
# LightRAG ATS - Production Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LightRAG
lightrag-hku>=1.0.0

# Database Drivers
asyncpg>=0.29.0
neo4j>=5.15.0
psycopg2-binary>=2.9.9

# Embedding & Reranking Models
sentence-transformers>=2.2.2
torch>=2.1.0

# Resume Parsing
pypdf>=3.17.0
python-docx>=1.1.0

# HTTP Client for Ollama
httpx>=0.26.0
aiohttp>=3.9.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
